Use supporter count query for ticker
This commit changes the value ticker to use a query that counts the
number of supporters (rather than the numeric value of support). This is
needed for the Australian banner that’s coming up in May.

@tomrf1 and I have tested in code that the state machine now reads the
correct data when called with the AUS ticker name. We can’t test the
ticker in code because SDC always queries the prod S3 files for ticker
data. Once this is merged we should be able to see the prod ticker data
in code by creating a ticker-containing banner in the code RRCP.